+++ client/contracts/BUILD.bazel
@@ -23,6 +23,7 @@ go_test(
     name = "go_default_test",
     srcs = ["sharding_manager_test.go"],
     embed = [":go_default_library"],
+    race = "off",
     deps = [
         "//client/params:go_default_library",
         "@com_github_ethereum_go_ethereum//accounts/abi/bind:go_default_library",
