+++ client/syncer/BUILD.bazel
@@ -45,6 +45,29 @@ go_test(
         "@com_github_ethereum_go_ethereum//core:go_default_library",
         "@com_github_ethereum_go_ethereum//core/types:go_default_library",
         "@com_github_ethereum_go_ethereum//crypto:go_default_library",
+        "@com_github_sirupsen_logrus//hooks/test:go_default_library",
+    ],
+)
+
+# by default gazelle tries to add all the test files to the  first
+# go_test; we need to exclude the tests that we want to configure
+# in a particular way
+# gazelle:exclude service_norace_test.go
+
+go_test(
+    name = "go_norace_test",
+    srcs = ["service_norace_test.go"],
+    embed = [":go_default_library"],
+    race = "off",  # TODO(#377): fix issues with race detection testing.
+    deps = [
+        "//client/mainchain:go_default_library",
+        "//client/params:go_default_library",
+        "//client/types:go_default_library",
+        "//proto/sharding/p2p/v1:go_default_library",
+        "//shared/database:go_default_library",
+        "//shared/p2p:go_default_library",
+        "@com_github_ethereum_go_ethereum//common:go_default_library",
+        "@com_github_ethereum_go_ethereum//core/types:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
         "@com_github_sirupsen_logrus//hooks/test:go_default_library",
     ],
