+++ shared/p2p/BUILD.bazel
@@ -33,7 +33,6 @@ go_library(
 go_test(
     name = "go_default_test",
     srcs = [
-        "discovery_test.go",
         "feed_example_test.go",
         "feed_test.go",
         "options_test.go",
@@ -47,22 +46,29 @@ go_test(
         "@com_github_ethereum_go_ethereum//event:go_default_library",
         "@com_github_golang_protobuf//proto:go_default_library",
         "@com_github_libp2p_go_floodsub//:go_default_library",
-        "@com_github_libp2p_go_libp2p//p2p/discovery:go_default_library",
         "@com_github_libp2p_go_libp2p//p2p/host/basic:go_default_library",
-        "@com_github_libp2p_go_libp2p_peer//:go_default_library",
         "@com_github_libp2p_go_libp2p_swarm//testing:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
-        "@com_github_sirupsen_logrus//hooks/test:go_default_library",
     ],
 )
 
 # by default gazelle tries to add all the test files to the  first
-# go_test; we want to treat feed_concurrent_test.go differently
-# gazelle:exclude feed_concurrent_test.go
+# go_test; we need to exclude the tests that we want to configure
+# in a particular way
+# gazelle:exclude discovery_norace_test.go
+# gazelle:exclude service_norace_test.go
 
 go_test(
-    name = "go_feed_concurrent_write_test",
-    srcs = ["feed_concurrent_test.go"],
+    name = "go_norace_test",
+    srcs = [
+        "discovery_norace_test.go",
+        "service_norace_test.go",
+    ],
     embed = [":go_default_library"],
-    race = "on",
+    race = "off",  # TODO(#377): fix issues with race detection testing.
+    deps = [
+        "@com_github_libp2p_go_libp2p//p2p/host/basic:go_default_library",
+        "@com_github_libp2p_go_libp2p_swarm//testing:go_default_library",
+        "@com_github_sirupsen_logrus//hooks/test:go_default_library",
+    ],
 )
