+++ shared/p2p/service_norace_test.go
@@ -1,0 +1,35 @@
+package p2p
+
+import (
+	"testing"
+
+	logTest "github.com/sirupsen/logrus/hooks/test"
+)
+
+func TestLifecycle(t *testing.T) {
+	hook := logTest.NewGlobal()
+
+	s, err := NewServer()
+	if err != nil {
+		t.Fatalf("Could not start a new server: %v", err)
+	}
+
+	s.Start()
+	msg := hook.Entries[0].Message
+	want := "Starting service"
+	if msg != want {
+		t.Errorf("incorrect log. wanted: %s. got: %v", want, msg)
+	}
+
+	s.Stop()
+	msg = hook.LastEntry().Message
+	want = "Stopping service"
+	if msg != want {
+		t.Errorf("incorrect log. wanted: %s. got: %v", want, msg)
+	}
+
+	// The context should have been cancelled.
+	if s.ctx.Err() == nil {
+		t.Error("Context was not cancelled")
+	}
+}
